Move I/O out of API file validation
