+++ api.js
@@ -1,12 +1,15 @@
 /**
  * @fileoverview Heart of the app, making it possible to call programmatically.
+ * Anything below this should not do any I/O (reads, writes, console, etc.).
  * Equivalent to cli.js in https://eslint.org/docs/developer-guide/architecture
  */
 
+const path = require("path");
+const xmlParser = require("./xmlParser");
 const reporter = require("./reporter");
 const folderParser = require("./folderParser");
 const pomParser = require("./pomParser");
-const validateApiFiles = require("./validateApiFiles");
+const validateApiFile = require("./validateApiFile");
 const validatePom = require("./validatePom");
 const validateGlobal = require("./validateGlobal");
 const validateImplementation = require("./validateImplementation");
@@ -14,6 +17,15 @@ const validateGitignore = require("./validateGitignore");
 const validateProperties = require("./validateProperties");
 const validateLog4j = require("./validateLog4j");
 const assert = require("./assert");
+
+function validateApiFiles(folderInfo, pomInfo) {
+  folderInfo.apiFiles.forEach(apiFile => {
+    let apiFileName = path.basename(apiFile);
+    let { xml } = xmlParser(apiFile);
+
+    validateApiFile(apiFileName, xml, pomInfo);
+  });
+}
 
 module.exports = {
   execute(apiBasePath) {
